pyproject.toml
[project]
name = "openai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-openai",
    "oci-openai",
    "python-dotenv",
]
